pyproject.toml
[tool.poetry]
name = 'lexicon'
version = '0.2.1'
description = 'Command line and lib toolset for word definitions, spell check, and spelling suggestions'
authors = ['roboto84 <7658704+roboto84@users.noreply.github.com>']

[tool.poetry.dependencies]
python = '^3.10'
requests = '^2.28.1'
python-dotenv = '^0.15.0'
pyenchant = '^3.2.0'
willow-core = {git = 'https://github.com/roboto84/willow_core.git', rev = '646a99f'}

[tool.poetry.dev-dependencies]
pytest = '^7.2.0'

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
